add jacobi symbol
quadratic reciprocity

+++ include/primes.hpp
@@ -27,8 +27,8 @@ namespace nm {
             ~Primality() {};
 
             bool fermat(std::uint64_t number);
-            bool miller_rabin(std::uint64_t number);
-            bool solovay_strassen(std::uint64_t number);
+            bool miller_rabin(std::uint32_t number);
+            bool solovay_strassen(std::uint32_t number);
             bool agarwal_kayal_saxena(std::uint64_t number);
     };
 } // primality tests
